bug fixed exception extension
bug fixed exception extension

+++ src/Delete.cpp
@@ -80,7 +80,7 @@ cpr::Response Delete::status_callback_ref(int& status_code, const cpr::Response&
    // dispatchEvent(*this);
 
 		//this->dispatchEvent();
-	} catch (rest::RestExceptionExt& e) {
+	} catch (rest::RestException& e) {
 		log_error("%s", e.what());
 	} catch (std::exception& e) {
 
@@ -91,23 +91,24 @@ cpr::Response Delete::status_callback_ref(int& status_code, const cpr::Response&
 
 void* Delete::run(void)
 {
-	log_info("Delete::run");
+	if (!this->is_running()) {
+		log_debug("Delete::run");
 
-	int m_status_code = 0;
-	auto callback = std::function<cpr::Response(cpr::Response)>(std::bind(Delete::status_callback_ref,
+		int m_status_code = 0;
+		auto callback = std::function<cpr::Response(cpr::Response)>(std::bind(Delete::status_callback_ref,
 			std::ref(m_status_code),
 			std::placeholders::_1));
 
-	log_debug("url: %s", m_url.c_str());
-	std::map<std::string, std::string>::iterator it = m_header.begin();
-	for(; it != m_header.end(); it++) {
-		log_debug ("Headers : \"%s\" : \"%s\"", it->first.c_str(), it->second.c_str());
+		log_debug("url: %s", m_url.c_str());
+		std::map<std::string, std::string>::iterator it = m_header.begin();
+		for (; it != m_header.end(); it++) {
+			log_debug("Headers : \"%s\" : \"%s\"", it->first.c_str(), it->second.c_str());
+		}
+		log_debug("m_body: %s", m_body.c_str());
+
+		auto future = cpr::DeleteCallback(callback, m_url, m_header, m_body, cpr::Timeout{ 1000 });
+		m_response = future.get();
 	}
-	log_debug("m_body: %s", m_body.c_str());
-
-	auto future = cpr::DeleteCallback(callback, m_url, m_header, m_body, cpr::Timeout{1000});
-	m_response = future.get();
-
 	return NULL;
 }
 
@@ -117,9 +118,6 @@ int Delete::send() throw (RestException, RestExceptionExt)
 
 	if(getResponseStatusCode() != 200)
 		throw (RestException(getResponseStatusCode()));
-
-	if(getResultCode() != 0 )
-		throw (RestExceptionExt(getResponseStatusCode(), getResultCode(), getResultMessage()));
 
 	return rtn;
 }
